+++ js/modules/base/colors.mjs
@@ -169,6 +169,7 @@ class ColorPalette {
 
 createRootColors();
 
-export { ColorPalette, getColor, findColor, addColor, adoptRootColors, getRootColors, extendRootColors, getRGBfromTColor, createRootColors, toHex };
+export { ColorPalette, getColor, findColor, addColor, adoptRootColors,
+         getRootColors, extendRootColors, getRGBfromTColor, createRootColors, toHex };
