+++ js/modules/gpad/TAxisPainter.mjs
@@ -3,28 +3,9 @@ import { select as d3_select, drag as d3_drag, timeFormat as d3_timeFormat,
          scaleTime as d3_scaleTime, scaleSymlog as d3_scaleSymlog,
          scaleLog as d3_scaleLog, scaleLinear as d3_scaleLinear } from '../d3.mjs';
 import { floatToString } from '../base/BasePainter.mjs';
-import { ObjectPainter } from '../base/ObjectPainter.mjs';
+import { ObjectPainter, EAxisBits } from '../base/ObjectPainter.mjs';
 import { FontHandler } from '../base/FontHandler.mjs';
 
-
-const EAxisBits = {
-   kDecimals: BIT(7),
-   kTickPlus: BIT(9),
-   kTickMinus: BIT(10),
-   kAxisRange: BIT(11),
-   kCenterTitle: BIT(12),
-   kCenterLabels: BIT(14),
-   kRotateTitle: BIT(15),
-   kPalette: BIT(16),
-   kNoExponent: BIT(17),
-   kLabelsHori: BIT(18),
-   kLabelsVert: BIT(19),
-   kLabelsDown: BIT(20),
-   kLabelsUp: BIT(21),
-   kIsInteger: BIT(22),
-   kMoreLogLabels: BIT(23),
-   kOppositeTitle: BIT(32) // atrificial bit, not possible to set in ROOT
-};
 
 /** @summary Return time offset value for given TAxis object
   * @private */
