+++ js/modules/gpad/TFramePainter.mjs
@@ -1,8 +1,8 @@
 import { gStyle, settings, isBatchMode, browser } from '../core.mjs';
 import { select as d3_select, pointer as d3_pointer, pointers as d3_pointers, drag as d3_drag } from '../d3.mjs';
-import { getActivePad, ObjectPainter } from '../base/ObjectPainter.mjs';
+import { getActivePad, ObjectPainter, EAxisBits } from '../base/ObjectPainter.mjs';
 import { getSvgLineStyle } from '../base/TAttLineHandler.mjs';
-import { EAxisBits, TAxisPainter } from './TAxisPainter.mjs';
+import { TAxisPainter } from './TAxisPainter.mjs';
 import { getElementRect, getAbsPosInCanvas } from '../base/BasePainter.mjs';
 import { FontHandler } from '../base/FontHandler.mjs';
 import { createMenu, closeMenu } from '../gui/menu.mjs';
