+++ js/modules/gpad/TPadPainter.mjs
@@ -1495,15 +1495,18 @@ class TPadPainter extends ObjectPainter {
       // find and remove painters which no longer exists in the list
       for (let k = 0; k < this.painters.length; ++k) {
          let sub = this.painters[k];
-         if ((sub.snapid === undefined) || sub.$secondary) continue; // look only for painters with snapid
+
+         if (typeof sub.snapid !== 'string') continue; // look only for painters with snapid
+
+         let snapid = sub.snapid, p = snapid.indexOf('#');
+         if (p > 0) snapid = snapid.slice(0, p);
 
          for (let i = 0; i < snap.fPrimitives.length; ++i)
-            if (snap.fPrimitives[i].fObjectID === sub.snapid) { sub = null; isanyfound = true; break; }
+            if (snap.fPrimitives[i].fObjectID === snapid) { sub = null; isanyfound = true; break; }
 
          if (sub) {
-            // console.log(`Remove painter ${k} from ${this.painters.length} class ${sub.getClassName()} ismain ${sub.isMainPainter()}`);
             // remove painter which does not found in the list of snaps
-            this.painters.splice(k--,1);
+            this.painters.splice(k--, 1);
             sub.cleanup(); // cleanup such painter
             isanyremove = true;
             if (this.main_painter_ref === sub)
