+++ js/modules/hist2d/TH2Painter.mjs
@@ -3,7 +3,7 @@ import { rgb as d3_rgb, chord as d3_chord, arc as d3_arc, ribbon as d3_ribbon } 
 import { TAttLineHandler } from '../base/TAttLineHandler.mjs';
 import { TAttMarkerHandler } from '../base/TAttMarkerHandler.mjs';
 import { TRandom, floatToString } from '../base/BasePainter.mjs';
-import { EAxisBits } from '../gpad/TAxisPainter.mjs';
+import { EAxisBits } from '../base/ObjectPainter.mjs';
 import { THistPainter } from './THistPainter.mjs';
 
 /**
