+++ js/modules/io.mjs
@@ -2923,7 +2923,9 @@ class TFile {
             if ((boundary[0] == '"') && (boundary[boundary.length - 1] == '"'))
                boundary = boundary.slice(1, boundary.length - 1);
             boundary = '--' + boundary;
-         } else console.error('Did not found boundary id in the response header');
+         } else {
+            console.error('Did not found boundary id in the response header');
+         }
 
          while (n < last) {
 
@@ -2934,7 +2936,8 @@ class TFile {
                code1 = code2;
                code2 = view.getUint8(o + 1);
 
-               if ((code1 == 13) && (code2 == 10)) {
+               if (((code1 == 13) && (code2 == 10)) || (code1 == 10)) {
+
                   if ((line.length > 2) && (line.slice(0, 2) == '--') && (line !== boundary))
                      return rejectFunc(Error(`Decode multipart message, expect boundary ${boundary} got ${line}`));
 
@@ -2955,8 +2958,10 @@ class TFile {
 
                   if ((nline > 1) && (line.length === 0)) finish_header = true;
 
-                  o++; nline++; line = '';
-                  code2 = view.getUint8(o + 1);
+                  nline++; line = '';
+                  if (code1 != 10) {
+                     o++; code2 = view.getUint8(o + 1);
+                  }
                } else {
                   line += String.fromCharCode(code1);
                }
